+++ src/main/java/me/playbosswar/com/gui/tasks/scheduler/MainScheduleMenu.java
@@ -62,7 +62,7 @@ public class MainScheduleMenu implements InventoryProvider {
                 e -> new EditDaysMenu(task).INVENTORY.open(player));
         contents.set(1, 3, clickableDaysItem);
 
-        contents.set(1, 7, ClickableItem.of(Items.getBackItem(), e -> new EditTaskMenu(task).INVENTORY.open(player)));
+        contents.set(2, 8, ClickableItem.of(Items.getBackItem(), e -> new EditTaskMenu(task).INVENTORY.open(player)));
     }
 
     @Override